controller: Update routes for config APIs
- Updates application ingress to route `<cluster>/api/*` to the config-api pod (was: `config.<cluster>/api/*`)
- Moves cortex alertmanager and ruler APIs into the config-api pod
- Updates the config-api pod with alertmanager and ruler endpoints to be proxied
- Refactors auth and proxy handling a bit for reuse across services
