+++ go/cmd/cortex/main.go
@@ -31,8 +31,6 @@ var (
 	listenAddress            string
 	cortexQuerierURL         string
 	cortexDistributorURL     string
-	cortexRulerURL           string
-	cortexAlertmanagerURL    string
 	tenantName               string
 	disableAPIAuthentication bool
 )
@@ -41,8 +39,6 @@ func main() {
 	flag.StringVar(&listenAddress, "listen", "", "")
 	flag.StringVar(&cortexQuerierURL, "cortex-querier-url", "", "")
 	flag.StringVar(&cortexDistributorURL, "cortex-distributor-url", "", "")
-	flag.StringVar(&cortexRulerURL, "cortex-ruler-url", "", "")
-	flag.StringVar(&cortexAlertmanagerURL, "cortex-alertmanager-url", "", "")
 	flag.StringVar(&tenantName, "tenantname", "", "")
 	flag.StringVar(&loglevel, "loglevel", "info", "error|info|debug")
 	flag.BoolVar(&disableAPIAuthentication, "disable-api-authn", false, "")
@@ -65,48 +61,31 @@ func main() {
 		log.Fatalf("bad cortex distributor URL: %s", uerr)
 	}
 
-	cortexrurl, uerr := url.Parse(cortexRulerURL)
-	if uerr != nil {
-		log.Fatalf("bad cortex ruler URL: %s", uerr)
-	}
-
-	cortexaurl, uerr := url.Parse(cortexAlertmanagerURL)
-	if uerr != nil {
-		log.Fatalf("bad cortex Alertmanager URL: %s", uerr)
-	}
-
 	if !disableAPIAuthentication {
 		middleware.ReadAuthTokenVerificationKeyFromEnvOrCrash()
 	}
 
 	log.Infof("cortex querier URL: %s", cortexqurl)
 	log.Infof("cortex distributor URL: %s", cortexdurl)
-	log.Infof("cortex ruler URL: %s", cortexrurl)
-	log.Infof("cortex Alertmanager URL: %s", cortexaurl)
 	log.Infof("listen address: %s", listenAddress)
 	log.Infof("tenant name: %s", tenantName)
 	log.Infof("API authentication enabled: %v", !disableAPIAuthentication)
 
-	reverseProxy := middleware.NewReverseProxy(tenantName, cortexqurl, cortexdurl,
-		cortexrurl, cortexaurl, disableAPIAuthentication)
+	// See: https://github.com/cortexproject/cortex/blob/master/docs/api/_index.md
+	cortexTenantHeader := "X-Scope-OrgID"
+	querierProxy := middleware.NewTenantReverseProxy(&tenantName, cortexTenantHeader, cortexqurl, disableAPIAuthentication)
+	distributorProxy := middleware.NewTenantReverseProxy(&tenantName, cortexTenantHeader, cortexdurl, disableAPIAuthentication)
 
 	// mux matches based on registration order, not prefix length.
 	router := mux.NewRouter()
 	// router.PathPrefix("/api/prom/push").HandlerFunc(reverseProxy.HandleWithDistributorProxy)
 
 	// Require non-deprecated push path (instead of also allowing /api/prom/push)
-	router.PathPrefix("/api/v1/push").HandlerFunc(reverseProxy.HandleWithDistributorProxy)
-
-	// Alertmanager and ruler
-	router.PathPrefix("/ruler").HandlerFunc(reverseProxy.HandleWithRulerProxy)
-	router.PathPrefix("/api/v1/rules").HandlerFunc(reverseProxy.HandleWithRulerProxy)
-	router.PathPrefix("/alertmanager").HandlerFunc(reverseProxy.HandleWithAlertmanagerProxy)
-	router.PathPrefix("/multitenant_alertmanager").HandlerFunc(reverseProxy.HandleWithAlertmanagerProxy)
-	router.PathPrefix("/api/v1/alerts").HandlerFunc(reverseProxy.HandleWithAlertmanagerProxy)
+	router.PathPrefix("/api/v1/push").HandlerFunc(distributorProxy.HandleWithProxy)
 
 	// /api/v1/read, /api/v1/query, /api/v1/labels etc: direct everything that's not
 	// /api/v1/push to the querier for now.
-	router.PathPrefix("/api/v1").HandlerFunc(reverseProxy.HandleWithQuerierProxy)
+	router.PathPrefix("/api/v1").HandlerFunc(querierProxy.HandleWithProxy)
 
 	router.Handle("/metrics", promhttp.Handler())
 	router.Use(middleware.PrometheusMetrics("cortex_api_proxy"))
